pyproject.toml
[tool.poetry]
name = "baseload"
version = "0.1.0"
description = ""
authors = ["Jan <abrell@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
streamlit = "^1.30.0"
plotly = "^5.18.0"
pyarrow = "^14.0.2"
scipy = "^1.12.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^7.0.6"


[tool.poetry.group.model.dependencies]
gamsapi = {extras = ["all"], version = ">=45.6.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
